Remove the Tribute row game: Reclamation is its redesign
Deletes the /tribute page, its components, engine, tests, devtools and
stylesheet, and the route and navbar link. The one helper Reclamation
borrowed, the element matchup blend, moves to expedition/elementMatchup.js.
405 tests green, vite build clean.
