+++ my-app/index.html
@@ -51,7 +51,6 @@
   <!-- Template Main CSS File -->
   <link href="/assets/css/style.css" rel="stylesheet">
   <link href="/assets/css/duel.css" rel="stylesheet">
-  <link href="/assets/css/tribute.css" rel="stylesheet">
   <link href="/assets/css/reclamation.css" rel="stylesheet">
   <link href="/assets/css/encyclopedia.css" rel="stylesheet">
   <link href="/assets/css/duel-playground.css" rel="stylesheet">
