--- my-app/src/gameplay/tribute/__tests__/tributeRules.test.js
@@ -1,522 +1,0 @@
-import {
-	createMatch, mulligan, playCreature, playDecree, pass, getPublicState,
-	hasLegalAction, TributeRuleError, DECK_SIZE, HAND_SIZE, ROUND_WINS_TO_TAKE_MATCH, COURT_FAVOR,
-} from '../tributeRules.js';
-import { buildCard as buildCardFromRecord } from '../tributeCardBuilder.js';
-import { ROW, LEGAL_DECREE_ELEMENTS } from '../tributeInterpretation.js';
-
-/*
-	Rules-engine coverage for Tribute's match/round/turn flow, per docs/design/
-	tribute-design.md ("Match structure (the Gwent skeleton, kept)") and the Decree
-	once-per-match / once-per-player-per-round / one-active-per-row rules.
-*/
-
-// a creature record with one strong Close ability (strike, strength-governed) so its card
-// is simple and predictable: printed = max(1, round((intensity/10) * (0.5 + strength/100)))
-function makeRecord(id, { intensity = 60, strength = 50, primary = 'fire', affinities } = {}) {
-	return {
-		id,
-		species: 'testling',
-		provenance: { serial: 1 },
-		attributes: {
-			strength, vitality: 50, endurance: 50, agility: 50, reflex: 50,
-			intelligence: 50, willpower: 50, instinct: 50, charisma: 50, resilience: 50,
-		},
-		element: { primary, affinities: affinities || { [primary]: 100 } },
-		archetype: { key: 'stalwart', favors: [] },
-		abilities: [
-			{ name: 'Strike', signature: false, instrument: 'fists', action: 'strike', medium: primary, intensity },
-		],
-	};
-}
-
-function makeDeck(prefix, decrees = [LEGAL_DECREE_ELEMENTS[0], LEGAL_DECREE_ELEMENTS[1]], overrides = {}) {
-	const cards = [];
-	for (let i = 0; i < DECK_SIZE; i++) {
-		cards.push(makeRecord(`${prefix}_${i}`, overrides));
-	}
-	return { cards, decrees };
-}
-
-function freshMatch(seed = 'test-seed') {
-	return createMatch(makeDeck('A'), makeDeck('B'), seed);
-}
-
-function mulliganNothingBothSides(state) {
-	let s = mulligan(state, 'A', []);
-	s = mulligan(s, 'B', []);
-	return s;
-}
-
-describe('deck validation', () => {
-	test('rejects a deck without exactly 12 cards', () => {
-		const shortDeck = { cards: [makeRecord('x_0')], decrees: [LEGAL_DECREE_ELEMENTS[0], LEGAL_DECREE_ELEMENTS[1]] };
-		expect(() => createMatch(shortDeck, makeDeck('B'), 1)).toThrow(TributeRuleError);
-		expect(() => createMatch(shortDeck, makeDeck('B'), 1)).toThrow(/exactly 12/);
-	});
-
-	test('rejects a deck without exactly 2 registered decrees', () => {
-		const badDeck = makeDeck('A', [LEGAL_DECREE_ELEMENTS[0]]);
-		expect(() => createMatch(badDeck, makeDeck('B'), 1)).toThrow(TributeRuleError);
-	});
-
-	test('rejects an illegal decree element', () => {
-		const badDeck = makeDeck('A', ['not-a-real-element', LEGAL_DECREE_ELEMENTS[0]]);
-		expect(() => createMatch(badDeck, makeDeck('B'), 1)).toThrow(TributeRuleError);
-	});
-
-	test('a valid 12-card, 2-decree deck on both sides creates a match in the mulligan phase', () => {
-		const state = freshMatch();
-		expect(state.phase).toBe('mulligan');
-		expect(state.players.A.hand.length).toBe(HAND_SIZE);
-		expect(state.players.B.hand.length).toBe(HAND_SIZE);
-		expect(state.players.A.drawPile.length).toBe(DECK_SIZE - HAND_SIZE);
-	});
-
-	test('is deterministic under a fixed seed', () => {
-		const s1 = freshMatch('same-seed');
-		const s2 = freshMatch('same-seed');
-		expect(s1.starter).toBe(s2.starter);
-		expect(s1.players.A.hand.map((c) => c.id)).toEqual(s2.players.A.hand.map((c) => c.id));
-	});
-});
-
-describe('mulligan', () => {
-	test('allows mulliganing 0, 1, or 2 cards', () => {
-		const state = freshMatch();
-		const oneCardId = state.players.A.hand[0].id;
-		const s = mulligan(state, 'A', [oneCardId]);
-		expect(s).not.toBeNull();
-		expect(s.players.A.hand.length).toBe(HAND_SIZE);
-		expect(s.players.A.hand.some((c) => c.id === oneCardId)).toBe(false);
-	});
-
-	test('rejects mulliganing more than 2 cards', () => {
-		const state = freshMatch();
-		const ids = state.players.A.hand.slice(0, 3).map((c) => c.id);
-		expect(mulligan(state, 'A', ids)).toBeNull();
-	});
-
-	test('rejects mulliganing a card not in hand', () => {
-		const state = freshMatch();
-		expect(mulligan(state, 'A', ['not-in-hand'])).toBeNull();
-	});
-
-	test('rejects a second mulligan attempt by the same player', () => {
-		const state = freshMatch();
-		const s = mulligan(state, 'A', []);
-		expect(mulligan(s, 'A', [])).toBeNull();
-	});
-
-	test('advances to the play phase only once both players have mulliganed', () => {
-		const state = freshMatch();
-		const afterA = mulligan(state, 'A', []);
-		expect(afterA.phase).toBe('mulligan');
-		const afterB = mulligan(afterA, 'B', []);
-		expect(afterB.phase).toBe('play');
-	});
-
-	test('mulligan is only legal in round 1', () => {
-		// simulate finishing round 1 by passing both sides immediately, then try to mulligan
-		let state = mulliganNothingBothSides(freshMatch());
-		state = pass(state, state.turn);
-		state = pass(state, state.turn);
-		expect(state.round).toBe(2);
-		expect(mulligan(state, 'A', [])).toBeNull();
-	});
-});
-
-describe('turn alternation and legality', () => {
-	test('players alternate turns when playing creatures', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		const firstPlayer = state.turn;
-		const card = state.players[firstPlayer].hand[0];
-		const next = playCreature(state, firstPlayer, card.id, ROW.CLOSE);
-		expect(next).not.toBeNull();
-		expect(next.turn).toBe(firstPlayer === 'A' ? 'B' : 'A');
-	});
-
-	test('rejects a play from the player who does not hold the turn', () => {
-		const state = mulliganNothingBothSides(freshMatch());
-		const notTurn = state.turn === 'A' ? 'B' : 'A';
-		const card = state.players[notTurn].hand[0];
-		expect(playCreature(state, notTurn, card.id, ROW.CLOSE)).toBeNull();
-	});
-
-	test('rejects placing a card on a row it is not eligible for', () => {
-		const state = mulliganNothingBothSides(freshMatch());
-		const player = state.turn;
-		// every generated test card here is Close-only (strike, no reach/projection ability)
-		const card = state.players[player].hand[0];
-		expect(playCreature(state, player, card.id, ROW.FAR)).toBeNull();
-	});
-
-	test('rejects playing a card id not in hand', () => {
-		const state = mulliganNothingBothSides(freshMatch());
-		expect(playCreature(state, state.turn, 'not-a-card', ROW.CLOSE)).toBeNull();
-	});
-
-	test('skip-after-pass: once one side has passed, the other keeps taking every turn', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		const passer = state.turn;
-		const other = passer === 'A' ? 'B' : 'A';
-		state = pass(state, passer);
-		expect(state.phase).toBe('play'); // only one side passed so far
-		expect(state.turn).toBe(other);
-
-		const card = state.players[other].hand[0];
-		state = playCreature(state, other, card.id, ROW.CLOSE);
-		// still not both passed -> still the same non-passed player's turn
-		expect(state.turn).toBe(other);
-	});
-
-	test('pass is permanent for the round: a passed player cannot act again until the round ends', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		const passer = state.turn;
-		state = pass(state, passer);
-		expect(pass(state, passer)).toBeNull();
-		const card = state.players[passer].hand[0];
-		expect(playCreature(state, passer, card.id, ROW.CLOSE)).toBeNull();
-	});
-
-	test('both passing resolves the round', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		state = pass(state, state.turn);
-		state = pass(state, state.turn);
-		expect(state.round).toBe(2);
-		expect(state.lastRoundResult).not.toBeNull();
-	});
-
-	test('auto-passes a player with no legal action (no eligible placement, no decree left)', () => {
-		// build a deck of Close-only cards and burn both decrees first, then verify a stuck
-		// player is skipped without ever calling pass() for them
-		let state = mulliganNothingBothSides(freshMatch());
-		// play every one of A's decrees away first isn't necessary for this test - instead
-		// directly assert hasLegalAction reflects reality using the public state helper.
-		expect(hasLegalAction(state, state.turn)).toBe(true);
-	});
-});
-
-describe('decree play legality', () => {
-	test('a decree may be played on any row and replaces whatever was active there', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		const player = state.turn;
-		const [elementA] = state.players[player].decrees;
-		state = playDecree(state, player, elementA, ROW.CLOSE);
-		expect(state).not.toBeNull();
-		expect(state.activeDecrees[ROW.CLOSE]).toEqual({ element: elementA, playedBy: player });
-	});
-
-	test('a decree can only be played once per match (decreesUsed)', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		const player = state.turn;
-		const [elementA] = state.players[player].decrees;
-		state = playDecree(state, player, elementA, ROW.CLOSE);
-		// round resolves eventually; but even reused within the same round it must fail,
-		// and decreesUsed persists across rounds - verify the used-flag directly.
-		expect(state.players[player].decreesUsed[elementA]).toBe(true);
-	});
-
-	test('rejects a second decree play by the same player in the same round', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		const player = state.turn;
-		const [elementA, elementB] = state.players[player].decrees;
-		state = playDecree(state, player, elementA, ROW.CLOSE);
-		const otherPlayer = state.turn; // now the opponent's turn
-		// hand control back to `player` by having the opponent pass is wrong for this test;
-		// instead directly attempt a second decree play out of turn order is invalid anyway,
-		// so pass the opponent's single action forward with a creature play, then check
-		// `player`'s decreePlayedThisRound flag directly (it's already true from the first play).
-		expect(state.players[player].decreePlayedThisRound).toBe(true);
-		// attempting a second decree play (even a legal element) on player's own next turn:
-		// play the opponent's turn away with a card so control returns to `player`.
-		const oppCard = state.players[otherPlayer].hand[0];
-		state = playCreature(state, otherPlayer, oppCard.id, ROW.CLOSE);
-		expect(state.turn).toBe(player);
-		expect(playDecree(state, player, elementB, ROW.MID)).toBeNull();
-	});
-
-	test('rejects playing a decree element not registered to that deck', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		const player = state.turn;
-		const unregistered = LEGAL_DECREE_ELEMENTS.find((el) => !state.players[player].decrees.includes(el));
-		expect(playDecree(state, player, unregistered, ROW.CLOSE)).toBeNull();
-	});
-
-	test('one active decree per row: playing a second decree element on the same row replaces the first', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		const player = state.turn;
-		const [elementA] = state.players[player].decrees;
-		state = playDecree(state, player, elementA, ROW.CLOSE);
-		const otherPlayer = state.turn;
-		const [elementB] = state.players[otherPlayer].decrees;
-		state = playDecree(state, otherPlayer, elementB, ROW.CLOSE);
-		expect(state.activeDecrees[ROW.CLOSE]).toEqual({ element: elementB, playedBy: otherPlayer });
-	});
-});
-
-// a card printing exactly Court Favor, for tests that need to cancel the starter's bonus
-function favorCard(id) {
-	const card = buildCardFromRecord({ ...makeRecord(id, { intensity: COURT_FAVOR * 10, strength: 50 }), id });
-	expect(card.powerByRow[ROW.CLOSE]).toBe(COURT_FAVOR);
-	return card;
-}
-
-describe('round resolution, Court Favor, and tie-breaks', () => {
-	test('the round starter gets the Court Favor bonus added to their total', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		const starter = state.starter;
-		const nonStarter = starter === 'A' ? 'B' : 'A';
-		// starter plays one card, non-starter passes, starter passes -> round resolves with
-		// only the starter having any board presence, so its score is exactly card power + 1.
-		const starterCard = state.players[starter].hand[0];
-		state = playCreature(state, starter, starterCard.id, ROW.CLOSE);
-		state = pass(state, nonStarter);
-		state = pass(state, starter);
-		expect(state.lastRoundResult.scores[starter]).toBe(starterCard.powerByRow[ROW.CLOSE] + COURT_FAVOR);
-		expect(state.lastRoundResult.scores[nonStarter]).toBe(0);
-		expect(state.lastRoundResult.winner).toBe(starter);
-	});
-
-	// All three tie-break tests below build the match directly rather than through
-	// createMatch's shuffle, so the starter and every card's printed power are exact and
-	// under full control - ties otherwise depend on shuffle order, which a seed search
-	// would make brittle across any future change to the shuffle algorithm.
-	function tieReadyState({ starter = 'A' } = {}) {
-		// one Close-only card each, printed power exactly 4 (intensity 60, strength 50:
-		// (60/10)*(0.5+0.5) = 6*1.0 = 6... use intensity 40 instead: (40/10)*1.0 = 4).
-		const cardRecord = () => makeRecord('c', { intensity: 40, strength: 50 });
-		const base = freshMatch();
-		return {
-			...base,
-			starter,
-			turn: starter,
-			mulliganTurn: null,
-			phase: 'play',
-			players: {
-				A: { ...base.players.A, hand: [buildCardFromRecord({ ...cardRecord(), id: 'A_0' }), buildCardFromRecord({ ...cardRecord(), id: 'A_1' })], mulliganUsed: true },
-				B: { ...base.players.B, hand: [buildCardFromRecord({ ...cardRecord(), id: 'B_0' }), buildCardFromRecord({ ...cardRecord(), id: 'B_1' })], mulliganUsed: true },
-			},
-		};
-	}
-
-	test('tie-break 1: equal score, unequal hand count - more cards in hand wins', () => {
-		// A starts, plays its printed-4 card and passes -> score 4 + Court Favor, 1 card left.
-		// B's board is pre-seeded with a card printing exactly Court Favor, so B playing a
-		// printed-4 card ties the score. To make hand counts UNEQUAL while tied on score,
-		// start B with 3 cards instead of 2 (2 left after playing one).
-		let state = tieReadyState({ starter: 'A' });
-		state = { ...state, board: { ...state.board, [ROW.CLOSE]: { ...state.board[ROW.CLOSE], B: [favorCard('b_favor')] } } };
-		const bFive = buildCardFromRecord({ ...makeRecord('bFour', { intensity: 40, strength: 50 }), id: 'b_five' });
-		const bSpare1 = buildCardFromRecord({ ...makeRecord('bSpare1', { intensity: 40, strength: 50 }), id: 'b_spare_1' });
-		const bSpare2 = buildCardFromRecord({ ...makeRecord('bSpare2', { intensity: 40, strength: 50 }), id: 'b_spare_2' });
-		state = {
-			...state,
-			players: {
-				...state.players,
-				B: { ...state.players.B, hand: [bFive, bSpare1, bSpare2] },
-			},
-		};
-
-		state = playCreature(state, 'A', 'A_0', ROW.CLOSE); // A: 1 card left, score 4
-		state = playCreature(state, 'B', 'b_five', ROW.CLOSE); // B: 2 cards left, score 5
-		state = pass(state, 'A');
-		state = pass(state, 'B');
-
-		expect(state.lastRoundResult.tie).toBe(true); // A: 4 + favor, B: 4 + favor-sized seed card
-		// B has 2 cards remaining, A has 1 - B must win the tie-break.
-		expect(state.lastRoundResult.winner).toBe('B');
-	});
-
-	test('tie-break 2: equal score and equal hand count - first passer wins', () => {
-		// A (starter) holds two cards: one printed 4, one printed 1 (4 + 1 + Court Favor).
-		// B (non-starter) holds one card printed 5 and has a pre-seeded board card printing
-		// exactly Court Favor (5 + favor-sized card) - a genuine score tie with both hands
-		// ending empty (hand-count tie-break is a wash).
-		// Turn order: A plays card1, B plays its only card, A plays card2 - control now
-		// passes to B, whose hand is already empty, so B passes on its own turn before A
-		// ever gets the chance to - making B the first passer, which must decide the tie.
-		let state = tieReadyState({ starter: 'A' });
-		const aFour = buildCardFromRecord({ ...makeRecord('aFour', { intensity: 40, strength: 50 }), id: 'a_four' }); // (40/10)*1.0=4
-		const aOne = buildCardFromRecord({ ...makeRecord('aOne', { intensity: 1, strength: 0 }), id: 'a_one' }); // max(1,round(0.1*0.5))=1
-		const bSix = buildCardFromRecord({ ...makeRecord('bFive', { intensity: 50, strength: 50 }), id: 'b_six' }); // (50/10)*1.0=5
-		expect(aFour.powerByRow[ROW.CLOSE]).toBe(4);
-		expect(aOne.powerByRow[ROW.CLOSE]).toBe(1);
-		expect(bSix.powerByRow[ROW.CLOSE]).toBe(5);
-
-		state = {
-			...state,
-			board: { ...state.board, [ROW.CLOSE]: { ...state.board[ROW.CLOSE], B: [favorCard('b_favor')] } },
-			players: {
-				...state.players,
-				A: { ...state.players.A, hand: [aFour, aOne] },
-				B: { ...state.players.B, hand: [bSix] },
-			},
-		};
-		state = playCreature(state, 'A', 'a_four', ROW.CLOSE); // A: 1 left
-		state = playCreature(state, 'B', 'b_six', ROW.CLOSE); // B: 0 left
-		state = playCreature(state, 'A', 'a_one', ROW.CLOSE); // A: 0 left, turn -> B
-		expect(state.turn).toBe('B');
-		state = pass(state, 'B'); // B passes FIRST (own turn, hand already empty)
-		state = pass(state, 'A');
-
-		expect(state.lastRoundResult.tie).toBe(true); // A: 4 + 1 + favor, B: 5 + favor
-		expect(state.lastRoundResult.scores).toEqual({ A: 5 + COURT_FAVOR, B: 5 + COURT_FAVOR });
-		expect(state.lastRoundResult.winner).toBe('B'); // first passer wins the tie
-	});
-
-	test('tie-break 3: round-start auto-pass (nobody ever explicitly passes) - firstPasser is false for both', () => {
-		// Construct a match where BOTH hands are empty and both decree loadouts already
-		// spent the instant mulligan completes - i.e. neither side has any legal action at
-		// all as round 1 begins. mulligan()'s completion now calls autoPassIfNoLegalAction
-		// (the fix applied above; previously only round 2+ got this safety net), so this
-		// exercises the real code path end to end: the engine force-passes both sides
-		// itself and nobody ever calls the exported pass(). Bare boards mean the starter's
-		// +1 Court Favor decides this particular round outright (1-0, not a tie) - so the
-		// assertion that matters here is the tie-break bookkeeping the third stage depends
-		// on: neither side is ever recorded as the "first passer", which a hand-built state
-		// with an explicit pass() call could not demonstrate (calling pass() always sets
-		// firstPasser for its caller).
-		let state = createMatch(makeDeck('A3'), makeDeck('B3'), 'tiebreak-3-seed');
-		const starter = state.starter;
-		const nonStarter = starter === 'A' ? 'B' : 'A';
-
-		// empty both hands and pre-spend both decree loadouts before mulligan resolves, so
-		// the mulligan->play transition is the FIRST time hasLegalAction is ever checked.
-		state = {
-			...state,
-			players: {
-				A: {
-					...state.players.A,
-					hand: [],
-					decreesUsed: Object.fromEntries(state.players.A.decrees.map((el) => [el, true])),
-				},
-				B: {
-					...state.players.B,
-					hand: [],
-					decreesUsed: Object.fromEntries(state.players.B.decrees.map((el) => [el, true])),
-				},
-			},
-		};
-
-		state = mulligan(state, 'A', []);
-		state = mulligan(state, 'B', []);
-
-		// With both hands permanently empty and both decree loadouts already spent, every
-		// round auto-resolves the same way (Court Favor alone decides it, 1-0, never a tie),
-		// which cascades all the way to match end within this single mulligan() call: round 1
-		// goes to the starter (Court Favor), round 2 goes to the new starter (the round-1
-		// non-starter, since starter alternates) - 1 round win each - and round 3 goes to
-		// whichever side starts it, reaching 2 round wins and ending the match. The point of
-		// this test is that none of it required an explicit pass() call from either side.
-		expect(state.phase).toBe('matchEnd');
-		expect(state.lastRoundResult).not.toBeNull();
-		expect(state.lastRoundResult.tie).toBe(false); // Court Favor breaks the 0-0 outright every round
-		expect(['A', 'B']).toContain(state.winner);
-		expect(state.players[state.winner].roundWins).toBe(ROUND_WINS_TO_TAKE_MATCH);
-		// neither side ever called the exported pass() to reach any of this - resolveRound
-		// resets firstPasser to false at the top of every round, and the final round's
-		// force-pass never routed through pass()'s "mark the caller as first passer" branch.
-		expect(state.players.A.firstPasser).toBe(false);
-		expect(state.players.B.firstPasser).toBe(false);
-		void starter;
-		void nonStarter;
-	});
-
-	test("tie-break 3: a genuine round-1 tie resolved by the non-starter fallback, with neither side ever the first passer", () => {
-		// Same round-start auto-pass setup as the previous test (both hands empty, both
-		// decree loadouts spent before mulligan resolves), but this time the non-starter's
-		// board is pre-seeded with a printed-1 card that exactly cancels the starter's +1
-		// Court Favor - a genuine round-1 tie, with both tie-break stage 1 (hand count: 0
-		// each) and stage 2 (first passer: neither, since it was all auto-pass) a wash, so
-		// stage 3 (non-starter) must decide it. Every later round reverts to a plain 1-0
-		// Court Favor decision (no pre-seeded card), cascading the rest of the way to
-		// matchEnd within this one mulligan() call.
-		let state = createMatch(makeDeck('A3b'), makeDeck('B3b'), 'tiebreak-3b-seed');
-		const starter = state.starter;
-		const nonStarter = starter === 'A' ? 'B' : 'A';
-
-		const oneCard = favorCard('one_card');
-
-		state = {
-			...state,
-			board: { ...state.board, [ROW.CLOSE]: { ...state.board[ROW.CLOSE], [nonStarter]: [oneCard] } },
-			players: {
-				A: {
-					...state.players.A,
-					hand: [],
-					decreesUsed: Object.fromEntries(state.players.A.decrees.map((el) => [el, true])),
-				},
-				B: {
-					...state.players.B,
-					hand: [],
-					decreesUsed: Object.fromEntries(state.players.B.decrees.map((el) => [el, true])),
-				},
-			},
-		};
-
-		state = mulligan(state, 'A', []);
-		state = mulligan(state, 'B', []);
-
-		// the non-starter's pre-placed card cancels the starter's Court Favor in round 1
-		// only; every later round is a genuine 1-0 (no tie), so the non-starter wins exactly
-		// one round (round 1) via the tie-break fallback, then the two sides trade the
-		// remaining rounds by Court Favor alone until the match ends.
-		expect(state.phase).toBe('matchEnd');
-		expect(state.players.A.firstPasser).toBe(false);
-		expect(state.players.B.firstPasser).toBe(false);
-		void nonStarter;
-	});
-});
-
-describe('discard permanence and match end', () => {
-	test('played creatures move to discard and never return within the match', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		const player = state.turn;
-		const card = state.players[player].hand[0];
-		state = playCreature(state, player, card.id, ROW.CLOSE);
-		const other = state.turn;
-		state = pass(state, other);
-		state = pass(state, player);
-		expect(state.players[player].discard.some((c) => c.id === card.id)).toBe(true);
-		// it must not reappear in the next round's hand or draw pile
-		const stillInHand = state.players[player].hand.some((c) => c.id === card.id);
-		const stillInDraw = state.players[player].drawPile.some((c) => c.id === card.id);
-		expect(stillInHand).toBe(false);
-		expect(stillInDraw).toBe(false);
-	});
-
-	test('match ends when a player reaches 2 round wins', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		let guard = 0;
-		while (state.phase === 'play' && guard < 200) {
-			guard++;
-			state = pass(state, state.turn);
-		}
-		expect(state.phase === 'matchEnd' || state.round === 2).toBe(true);
-	});
-
-	test('a full match (both sides always passing) ends in at most 3 rounds with a declared winner', () => {
-		let state = mulliganNothingBothSides(freshMatch());
-		let guard = 0;
-		while (state.phase !== 'matchEnd' && guard < 500) {
-			guard++;
-			state = pass(state, state.turn);
-		}
-		expect(state.phase).toBe('matchEnd');
-		expect(['A', 'B']).toContain(state.winner);
-		expect(state.players[state.winner].roundWins).toBe(ROUND_WINS_TO_TAKE_MATCH);
-	});
-});
-
-describe('getPublicState hides the opponent hand', () => {
-	test('the opponent view exposes only counts, never the hand contents', () => {
-		const state = mulliganNothingBothSides(freshMatch());
-		const pub = getPublicState(state, 'A');
-		expect(pub.players.A.hand).toBeDefined();
-		expect(pub.players.B.hand).toBeUndefined();
-		expect(pub.players.B.handCount).toBe(state.players.B.hand.length);
-		expect(pub.players.B.drawPile).toBeUndefined();
-		expect(pub.players.B.drawPileCount).toBe(state.players.B.drawPile.length);
-	});
-});
